refactor: remove csi-common package, move server.go into azurelustre
The csi-common package was a copy of upstream CSI utility code that
we have gradually replaced. After removing the driver, default server,
and identity abstractions in prior commits, only server.go and
utils.go remain. Move server.go into the azurelustre package (its
only consumer) and delete utils.go (unused after earlier cleanups).
